Rename select-fields to field-selection.ts, expose path reconstruction
